implement polygon winding tests when marshaling
RFC-7946 stipulates that external boundaries of Polygons should be
counter-clockwise, and that internal boundaries should be clockwise.
This commit adds functionality to test the winding orientation of
coordinate rings, and reverses coordinate direction when the incorrent
orientation is found while marshaling.

There are no changes to unmarshaling, the coordinate order is returned
as it is read.